doc-rag-portfolio/requirements.txt
# Document RAG + long-form generation — portfolio project
# Python 3.10+

# PDF
pypdf>=3.0.0

# Embeddings + vector store (local, no cloud signup required)
chromadb>=0.4.22
openai>=1.12.0

# App
streamlit>=1.29.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0
